# docs/swagger.yaml
basePath: /
definitions:
  ds.Device:
    properties:
      avgMaxPower:
        type: number
      avgMinPower:
        type: number
      id:
        type: integer
      image:
        type: string
      maxRadiationZone:
        type: string
      maxSafeRange:
        type: number
      minSafeRange:
        type: number
      radiationSource:
        type: string
      radiationType:
        type: string
      title:
        type: string
      visability:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API сервиса аутентификации с Redis и JWT. Поддерживает роли пользователей
    и администраторов.
  title: Device emission Service API
  version: "1.0"
paths:
  /api/device:
    post:
      consumes:
      - application/json
      description: Создает новое устройство с указанными параметрами. Доступно только
        администраторам
      parameters:
      - description: Данные нового устройства
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.Device'
      produces:
      - application/json
      responses:
        "201":
          description: Устройство успешно создано
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректные данные устройства
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (требуется роль Admin)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при создании устройства
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать новое устройство
      tags:
      - Устройства
  /api/device/:id:
    get:
      description: Возвращает информацию об устройстве по его идентификатору
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с данными устройства
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID устройства
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка на сервере
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить устройство по ID
      tags:
      - Устройства
    put:
      consumes:
      - application/json
      description: Обновляет данные существующего устройства по его ID. Доступно только
        администраторам
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные устройства
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.Device'
      produces:
      - application/json
      responses:
        "200":
          description: Устройство успешно обновлено
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID или данные устройства
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (требуется роль Admin)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при обновлении устройства
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить устройство
      tags:
      - Устройства
  /api/device/:id/delete:
    post:
      description: Помечает устройство как скрытое (soft delete). Доступно только
        администраторам
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Устройство успешно скрыто
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный ID устройства
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (требуется роль Admin)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при удалении устройства
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить устройство
      tags:
      - Устройства
  /api/device/:id/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает или заменяет изображение для устройства. Доступно только
        администраторам
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      - description: Изображение устройства
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Изображение успешно загружено
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный ID или отсутствует файл изображения
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (требуется роль Admin)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при загрузке изображения
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Загрузить изображение устройства
      tags:
      - Устройства
  /api/devices:
    get:
      description: Возвращает список всех устройств с возможностью фильтрации по названию
      parameters:
      - description: Фильтр по названию устройства
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком устройств
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка на сервере
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список устройств
      tags:
      - Устройства
...
  /api/emissions_calculation/draft/add/:id:
    post:
      description: Добавляет устройство в черновой заказ пользователя. Если чернового
        заказа нет, создает новый
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Устройство успешно добавлено в черновой заказ
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID устройства
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при добавлении устройства
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить устройство в черновой заказ
      tags:
      - Заявки с устройствами
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"